refactor: move dcejson exporter unified cli
